apps/api/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioredis==2.0.1
APScheduler==3.10.4
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-asgi-testclient==1.4.11
async-timeout==4.0.3
asyncmy==0.2.9
asyncpg==0.30.0
attrs==25.1.0
bcrypt==4.1.2
black==24.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
docker==7.1.0
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
numpy==2.0.2
packaging==24.0
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
propcache==0.2.1
psutil==6.1.1
psycopg2-binary==2.9.10
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
pytest==8.1.1
pytest-alembic==0.11.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.1
redis==5.0.4
requests==2.31.0
rsa==4.9
ruff==0.3.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2025.1
urllib3==1.26.16
uvicorn==0.29.0
yarl==1.18.3
